Implement access delegation using JWT delegation tokens
When authorizations are enabled some people might not be
able to log in a given container because of a lack of
permissions. In that case, a super admin user can delegate
the rights for this user to log into a specific container
identified by a task_id for a certain amount of time.

The documentation on how to produce such tokens if you are
a super admin is available by doing a GET query on /delegate

The feature is disabled by default but can be activated
with the ENABLE_RIGHTS_DELEGATION env variable set to true.