Require authentication again when first attempt failed. (#4)
* Require authentication again when first attempt failed.

In firefox, upon bad authentication (with wrong credentials),
MesosTerm was then always returning a 401 error with a page containing the
message "Unauthorized". From this time on, ao authentication request was pushed
to the user anymore and therefore there was no way for him/her to fix the
credentials. The only workaround was to clear the browser cache.

* Remove bad tsc dependency.