Allow tests to be launched from docker
Having a working setup of nodejs on my machine is painful, especially
considering that tests must run with nodejs 11 (I don't know exactly
why).
This patch makes it possible for a developer to run tests from a
container

A future patch could modify the way travis runs tests to use this
container.

Change-Id: Ic2ce5adb7d7db58cac3b55ca9a11ce1fd8ebe7f3