Detect EOF in FileReader when container dies.
Previously, the file reader was checking for new available data even
when the container was dead.

This lead to a side effect when the task disappeared from the mesos
state. Indeed, if the browser was open for long enough, the file reader
was still checking every 2s if new data was available even when the
container was dead and until the container disappears from the mesos
state. When the container disappeared from mesos state, some side
effects unveiled.

Since the regular check (2s) was querying the mesos state cache to find
the slave where the task was running and since the task disappeared
from mesos state cache, a cache invalidation was always triggered every
2s.
This is explained by the fact that the mesos state cache is invalidated
when a query for a task id is done and the task does not exist in cache
yet. This feature allows to refresh the cache before the next regular
refresh to improve the user experience. It allows to not show an error
message when the container has just spawned.

The fix in this PR detects when the task is dead and forwards that
information to the file reader which can then stop listening for new
updates.
Please note that the container status information has the same
refresh rate as the mesos state cache. So even when the container dies
the file reader will keep checking new updates until next mesos state
refresh.