Find all tasks in mesos agent state
Before this patch, some tasks were marked as not found when trying to
explore dead tasks sandbox.
This was due to agent state being organized between `frameworks` and
`completed_frameworks`.
We now look for task info in all parts of the state and thus allow to
explore all dead tasks sandbox.

Change-Id: I20e675fd37ece6a37b6273134785ed5537c0cf49
JIRA: MESOS-4717